+++ src/components/Card/index.ts
@@ -1,1 +1,3 @@
+// Card from copy-ui @ 2025-10-31.
+
 export { Card } from "./card";