+++ src/components/Card/CHANGELOG.md
@@ -1,1 +1,3 @@
 - 2025-10-21: Initial version.
+- 2025-10-31: Add the sub-component `Card.Footer`, update styles to make it
+  looks better, etc.